git: check-attr
